move getTopLevelElementByExportsDetail
