+++ editor/src/core/shared/project-file-types.ts
@@ -765,14 +765,3 @@ export type NodeModules = {
 
 // Key here is the full filename.
 export type ProjectContents = { [filepath: string]: ProjectFile }
-
-export function getTopLevelElementByExportsDetail(
-  file: ParseSuccess,
-  nameToLookFor: string,
-): UtopiaJSXComponent | null {
-  return (
-    file.topLevelElements.find(
-      (t): t is UtopiaJSXComponent => t.type === 'UTOPIA_JSX_COMPONENT' && t.name === nameToLookFor,
-    ) ?? null
-  )
-}
